Move test-common to src/main
