+++ engine/src/test/java/com/google/android/fhir/impl/FhirEngineImplTest.kt
@@ -24,10 +24,12 @@ import com.google.android.fhir.db.ResourceNotFoundException
 import com.google.android.fhir.db.impl.dao.LocalChangeToken
 import com.google.android.fhir.get
 import com.google.android.fhir.logicalId
-import com.google.android.fhir.resource.TestingUtils
 import com.google.android.fhir.search.search
 import com.google.android.fhir.sync.AcceptLocalConflictResolver
 import com.google.android.fhir.sync.AcceptRemoteConflictResolver
+import com.google.android.fhir.testing.assertResourceEquals
+import com.google.android.fhir.testing.assertResourceNotEquals
+import com.google.android.fhir.testing.readFromFile
 import com.google.common.truth.Truth.assertThat
 import java.util.Date
 import kotlinx.coroutines.Dispatchers
@@ -52,7 +54,6 @@ import org.robolectric.RobolectricTestRunner
 class FhirEngineImplTest {
   private val services = builder(ApplicationProvider.getApplicationContext()).inMemory().build()
   private val fhirEngine = services.fhirEngine
-  private val testingUtils = TestingUtils(services.parser)
 
   @Before fun setUp(): Unit = runBlocking { fhirEngine.create(TEST_PATIENT_1) }
 
@@ -60,15 +61,15 @@ class FhirEngineImplTest {
   fun create_shouldCreateResource() = runBlocking {
     val ids = fhirEngine.create(TEST_PATIENT_2)
     assertThat(ids).containsExactly("test_patient_2")
-    testingUtils.assertResourceEquals(TEST_PATIENT_2, fhirEngine.get<Patient>(TEST_PATIENT_2_ID))
+    assertResourceEquals(TEST_PATIENT_2, fhirEngine.get<Patient>(TEST_PATIENT_2_ID))
   }
 
   @Test
   fun createAll_shouldCreateResource() = runBlocking {
     val ids = fhirEngine.create(TEST_PATIENT_1, TEST_PATIENT_2)
     assertThat(ids).containsExactly("test_patient_1", "test_patient_2")
-    testingUtils.assertResourceEquals(TEST_PATIENT_1, fhirEngine.get<Patient>(TEST_PATIENT_1_ID))
-    testingUtils.assertResourceEquals(TEST_PATIENT_2, fhirEngine.get<Patient>(TEST_PATIENT_2_ID))
+    assertResourceEquals(TEST_PATIENT_1, fhirEngine.get<Patient>(TEST_PATIENT_1_ID))
+    assertResourceEquals(TEST_PATIENT_2, fhirEngine.get<Patient>(TEST_PATIENT_2_ID))
   }
 
   @Test
@@ -84,10 +85,7 @@ class FhirEngineImplTest {
       }
     val ids = fhirEngine.create(patient.copy())
     assertThat(ids).hasSize(1)
-    testingUtils.assertResourceEquals(
-      patient.setId(ids.first()),
-      fhirEngine.get<Patient>(ids.first())
-    )
+    assertResourceEquals(patient.setId(ids.first()), fhirEngine.get<Patient>(ids.first()))
   }
 
   @Test
@@ -120,14 +118,8 @@ class FhirEngineImplTest {
 
     fhirEngine.update(updatedPatient1, updatedPatient2)
 
-    testingUtils.assertResourceEquals(
-      updatedPatient1,
-      fhirEngine.get<Patient>("test-update-patient-001")
-    )
-    testingUtils.assertResourceEquals(
-      updatedPatient2,
-      fhirEngine.get<Patient>("test-update-patient-002")
-    )
+    assertResourceEquals(updatedPatient1, fhirEngine.get<Patient>("test-update-patient-001"))
+    assertResourceEquals(updatedPatient2, fhirEngine.get<Patient>("test-update-patient-002"))
   }
 
   @Test
@@ -151,11 +143,8 @@ class FhirEngineImplTest {
       runBlocking { fhirEngine.update(updatedPatient1, updatedPatient2) }
     }
 
-    testingUtils.assertResourceNotEquals(
-      updatedPatient1,
-      fhirEngine.get<Patient>("test-update-patient-001")
-    )
-    testingUtils.assertResourceEquals(patient1, fhirEngine.get<Patient>("test-update-patient-001"))
+    assertResourceNotEquals(updatedPatient1, fhirEngine.get<Patient>("test-update-patient-001"))
+    assertResourceEquals(patient1, fhirEngine.get<Patient>("test-update-patient-001"))
   }
 
   @Test
@@ -172,7 +161,7 @@ class FhirEngineImplTest {
 
   @Test
   fun load_shouldReturnResource() = runBlocking {
-    testingUtils.assertResourceEquals(TEST_PATIENT_1, fhirEngine.get<Patient>(TEST_PATIENT_1_ID))
+    assertResourceEquals(TEST_PATIENT_1, fhirEngine.get<Patient>(TEST_PATIENT_1_ID))
   }
 
   @Test
@@ -277,7 +266,7 @@ class FhirEngineImplTest {
   fun syncDownload_downloadResources() = runBlocking {
     fhirEngine.syncDownload(AcceptLocalConflictResolver) { flowOf((listOf((TEST_PATIENT_2)))) }
 
-    testingUtils.assertResourceEquals(TEST_PATIENT_2, fhirEngine.get<Patient>(TEST_PATIENT_2_ID))
+    assertResourceEquals(TEST_PATIENT_2, fhirEngine.get<Patient>(TEST_PATIENT_2_ID))
   }
 
   private fun buildPatient(
@@ -294,7 +283,7 @@ class FhirEngineImplTest {
 
   @Test
   fun `getLocalChange() should return single local change`() = runBlocking {
-    val patient: Patient = testingUtils.readFromFile(Patient::class.java, "/date_test_patient.json")
+    val patient: Patient = readFromFile(Patient::class.java, "/date_test_patient.json")
     fhirEngine.create(patient)
     val patientString = services.parser.encodeResourceToString(patient)
     val squashedLocalChange = fhirEngine.getLocalChange(patient.resourceType, patient.logicalId)
@@ -308,7 +297,7 @@ class FhirEngineImplTest {
 
   @Test
   fun `getLocalChange() should return squashed local change`() = runBlocking {
-    val patient: Patient = testingUtils.readFromFile(Patient::class.java, "/date_test_patient.json")
+    val patient: Patient = readFromFile(Patient::class.java, "/date_test_patient.json")
     fhirEngine.create(patient)
 
     patient.gender = Enumerations.AdministrativeGender.FEMALE
@@ -328,14 +317,14 @@ class FhirEngineImplTest {
 
   @Test
   fun `getLocalChange() with wrong resource id should return null`() = runBlocking {
-    val patient: Patient = testingUtils.readFromFile(Patient::class.java, "/date_test_patient.json")
+    val patient: Patient = readFromFile(Patient::class.java, "/date_test_patient.json")
     fhirEngine.create(patient)
     assertThat(fhirEngine.getLocalChange(patient.resourceType, "nonexistent_patient")).isNull()
   }
 
   @Test
   fun `getLocalChange() with wrong resource type should return null`() = runBlocking {
-    val patient: Patient = testingUtils.readFromFile(Patient::class.java, "/date_test_patient.json")
+    val patient: Patient = readFromFile(Patient::class.java, "/date_test_patient.json")
     fhirEngine.create(patient)
 
     assertThat(fhirEngine.getLocalChange(ResourceType.Encounter, patient.logicalId)).isNull()
@@ -343,7 +332,7 @@ class FhirEngineImplTest {
 
   @Test
   fun `clearDatabase() should clear all tables data`() = runBlocking {
-    val patient: Patient = testingUtils.readFromFile(Patient::class.java, "/date_test_patient.json")
+    val patient: Patient = readFromFile(Patient::class.java, "/date_test_patient.json")
     fhirEngine.create(patient)
     val patientString = services.parser.encodeResourceToString(patient)
     val squashedLocalChange = fhirEngine.getLocalChange(patient.resourceType, patient.logicalId)
@@ -353,10 +342,7 @@ class FhirEngineImplTest {
       assertThat(type).isEqualTo(Type.INSERT)
       assertThat(payload).isEqualTo(patientString)
     }
-    testingUtils.assertResourceEquals(
-      patient,
-      fhirEngine.get(ResourceType.Patient, patient.logicalId)
-    )
+    assertResourceEquals(patient, fhirEngine.get(ResourceType.Patient, patient.logicalId))
     // clear databse
     runBlocking(Dispatchers.IO) { fhirEngine.clearDatabase() }
     // assert that previously present resource not available after clearing database
@@ -386,16 +372,16 @@ class FhirEngineImplTest {
 
   @Test
   fun `purge() with local change and force purge false should throw IllegalStateException`() =
-      runBlocking {
-    val resourceIllegalStateException =
-      assertThrows(IllegalStateException::class.java) {
-        runBlocking { fhirEngine.purge(ResourceType.Patient, TEST_PATIENT_1_ID) }
-      }
-    assertThat(resourceIllegalStateException.message)
-      .isEqualTo(
-        "Resource with type ${TEST_PATIENT_1.resourceType.name} and id $TEST_PATIENT_1_ID has local changes, either sync with server or FORCE_PURGE required"
-      )
-  }
+    runBlocking {
+      val resourceIllegalStateException =
+        assertThrows(IllegalStateException::class.java) {
+          runBlocking { fhirEngine.purge(ResourceType.Patient, TEST_PATIENT_1_ID) }
+        }
+      assertThat(resourceIllegalStateException.message)
+        .isEqualTo(
+          "Resource with type ${TEST_PATIENT_1.resourceType.name} and id $TEST_PATIENT_1_ID has local changes, either sync with server or FORCE_PURGE required"
+        )
+    }
 
   @Test
   fun `purge() resource not available should throw ResourceNotFoundException`() = runBlocking {
@@ -448,68 +434,66 @@ class FhirEngineImplTest {
         }
       )
       .isEmpty()
-    testingUtils.assertResourceEquals(fhirEngine.get<Patient>("original-001"), remoteChange)
+    assertResourceEquals(fhirEngine.get<Patient>("original-001"), remoteChange)
   }
 
   @Test
   fun syncDownload_conflictResolution_acceptLocal_shouldHaveLocalChangeCreatedAgainstRemoteVersion() =
-      runBlocking {
-    val originalPatient =
-      Patient().apply {
-        id = "original-002"
-        meta =
-          Meta().apply {
-            versionId = "1"
-            lastUpdated = Date()
-          }
-        addName(
-          HumanName().apply {
-            family = "Stark"
-            addGiven("Tony")
-          }
+    runBlocking {
+      val originalPatient =
+        Patient().apply {
+          id = "original-002"
+          meta =
+            Meta().apply {
+              versionId = "1"
+              lastUpdated = Date()
+            }
+          addName(
+            HumanName().apply {
+              family = "Stark"
+              addGiven("Tony")
+            }
+          )
+        }
+      fhirEngine.syncDownload(AcceptLocalConflictResolver) { flowOf((listOf((originalPatient)))) }
+      var localChange =
+        originalPatient.copy().apply { addAddress(Address().apply { city = "Malibu" }) }
+      fhirEngine.update(localChange)
+
+      localChange =
+        localChange.copy().apply {
+          addAddress(
+            Address().apply {
+              city = "Malibu"
+              state = "California"
+            }
+          )
+        }
+      fhirEngine.update(localChange)
+
+      val remoteChange =
+        originalPatient.copy().apply {
+          meta =
+            Meta().apply {
+              versionId = "2"
+              lastUpdated = Date()
+            }
+          addAddress(Address().apply { country = "USA" })
+        }
+
+      fhirEngine.syncDownload(AcceptLocalConflictResolver) { flowOf((listOf(remoteChange))) }
+
+      val localChangeDiff =
+        """[{"op":"remove","path":"\/address\/0\/country"},{"op":"add","path":"\/address\/0\/city","value":"Malibu"},{"op":"add","path":"\/address\/-","value":{"city":"Malibu","state":"California"}}]"""
+      assertThat(
+          services.database
+            .getAllLocalChanges()
+            .first { it.localChange.resourceId == "original-002" }
+            .localChange.payload
         )
-      }
-    fhirEngine.syncDownload(AcceptLocalConflictResolver) { flowOf((listOf((originalPatient)))) }
-    var localChange =
-      originalPatient.copy().apply { addAddress(Address().apply { city = "Malibu" }) }
-    fhirEngine.update(localChange)
-
-    localChange =
-      localChange.copy().apply {
-        addAddress(
-          Address().apply {
-            city = "Malibu"
-            state = "California"
-          }
-        )
-      }
-    fhirEngine.update(localChange)
-
-    val remoteChange =
-      originalPatient.copy().apply {
-        meta =
-          Meta().apply {
-            versionId = "2"
-            lastUpdated = Date()
-          }
-        addAddress(Address().apply { country = "USA" })
-      }
-
-    fhirEngine.syncDownload(AcceptLocalConflictResolver) { flowOf((listOf(remoteChange))) }
-
-    val localChangeDiff =
-      """[{"op":"remove","path":"\/address\/0\/country"},{"op":"add","path":"\/address\/0\/city","value":"Malibu"},{"op":"add","path":"\/address\/-","value":{"city":"Malibu","state":"California"}}]"""
-    assertThat(
-        services
-          .database
-          .getAllLocalChanges()
-          .first { it.localChange.resourceId == "original-002" }
-          .localChange
-          .payload
-      )
-      .isEqualTo(localChangeDiff)
-    testingUtils.assertResourceEquals(fhirEngine.get<Patient>("original-002"), localChange)
-  }
+        .isEqualTo(localChangeDiff)
+      assertResourceEquals(fhirEngine.get<Patient>("original-002"), localChange)
+    }
 
   companion object {
     private const val TEST_PATIENT_1_ID = "test_patient_1"
