+++ engine/src/test/java/com/google/android/fhir/sync/AcceptRemoteConflictResolverTest.kt
@@ -16,9 +16,7 @@
 
 package com.google.android.fhir.sync
 
-import ca.uhn.fhir.context.FhirContext
-import ca.uhn.fhir.context.FhirVersionEnum
-import com.google.android.fhir.resource.TestingUtils
+import com.google.android.fhir.testing.assertResourceEquals
 import com.google.common.truth.Truth.assertThat
 import org.hl7.fhir.r4.model.HumanName
 import org.hl7.fhir.r4.model.Patient
@@ -28,8 +26,6 @@ import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
 class AcceptRemoteConflictResolverTest {
-  private val testingUtils = TestingUtils(FhirContext.forCached(FhirVersionEnum.R4).newJsonParser())
-
   @Test
   fun resolve_shouldReturnRemoteChange() {
     val localResource =
@@ -56,6 +52,6 @@ class AcceptRemoteConflictResolverTest {
 
     val result = AcceptRemoteConflictResolver.resolve(localResource, remoteResource)
     assertThat(result).isInstanceOf(Resolved::class.java)
-    testingUtils.assertResourceEquals(remoteResource, (result as Resolved).resolved)
+    assertResourceEquals(remoteResource, (result as Resolved).resolved)
   }
 }
