+++ engine/src/test/java/com/google/android/fhir/sync/FhirSyncWorkerTest.kt
@@ -23,7 +23,10 @@ import androidx.work.ListenableWorker
 import androidx.work.WorkerParameters
 import androidx.work.testing.TestListenableWorkerBuilder
 import com.google.android.fhir.FhirEngine
-import com.google.android.fhir.resource.TestingUtils
+import com.google.android.fhir.testing.TestDataSourceImpl
+import com.google.android.fhir.testing.TestDownloadManagerImpl
+import com.google.android.fhir.testing.TestFailingDatasource
+import com.google.android.fhir.testing.TestFhirEngineImpl
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
@@ -39,20 +42,18 @@ class FhirSyncWorkerTest {
   class PassingPeriodicSyncWorker(appContext: Context, workerParams: WorkerParameters) :
     FhirSyncWorker(appContext, workerParams) {
 
-    override fun getFhirEngine(): FhirEngine = TestingUtils.TestFhirEngineImpl
-    override fun getDataSource(): DataSource = TestingUtils.TestDataSourceImpl
-    override fun getDownloadWorkManager(): DownloadWorkManager =
-      TestingUtils.TestDownloadManagerImpl()
+    override fun getFhirEngine(): FhirEngine = TestFhirEngineImpl
+    override fun getDataSource(): DataSource = TestDataSourceImpl
+    override fun getDownloadWorkManager(): DownloadWorkManager = TestDownloadManagerImpl()
     override fun getConflictResolver() = AcceptRemoteConflictResolver
   }
 
   class FailingPeriodicSyncWorker(appContext: Context, workerParams: WorkerParameters) :
     FhirSyncWorker(appContext, workerParams) {
 
-    override fun getFhirEngine(): FhirEngine = TestingUtils.TestFhirEngineImpl
-    override fun getDataSource(): DataSource = TestingUtils.TestFailingDatasource
-    override fun getDownloadWorkManager(): DownloadWorkManager =
-      TestingUtils.TestDownloadManagerImpl()
+    override fun getFhirEngine(): FhirEngine = TestFhirEngineImpl
+    override fun getDataSource(): DataSource = TestFailingDatasource
+    override fun getDownloadWorkManager(): DownloadWorkManager = TestDownloadManagerImpl()
     override fun getConflictResolver() = AcceptRemoteConflictResolver
   }
 
@@ -61,8 +62,8 @@ class FhirSyncWorkerTest {
     workerParams: WorkerParameters
   ) : FhirSyncWorker(appContext, workerParams) {
 
-    override fun getFhirEngine(): FhirEngine = TestingUtils.TestFhirEngineImpl
-    override fun getDownloadWorkManager() = TestingUtils.TestDownloadManagerImpl()
+    override fun getFhirEngine(): FhirEngine = TestFhirEngineImpl
+    override fun getDownloadWorkManager() = TestDownloadManagerImpl()
     override fun getDataSource(): DataSource? = null
     override fun getConflictResolver() = AcceptRemoteConflictResolver
   }
